Create ArduinoSketchCode
This code is required for taking the reading of the from the sensors attached to the patients body. This also includes the code for sending the reading to web service ThingSpeak where the readings are displayed.  This can help doctors to remotely observe patient health conditions and also prescribe patients in an in person visit is not necessary.
Sensors measure body temperature and pulse rate of the patient.